Simplify and clarify game issue template
Debating whether or not an "OS" line should be added under CPU, GPU and RAM.

Removed "current on master" as it was redundant with "emulator build version", and the latter is more understandable for most, though I'm open to keeping "current on master". 

Moved "hardware specs" up for better readability.

Removed "pull requests needed" as nobody uses that anymore (not as a different section at least, no point in doing that) and added it to the game state description.

Removed "outstanding issues" as we're no longer taking game-specific issues on the main repository.

Expanded on "describe the current state of the game" to make up for the above change.

Specified that game screenshots are preferable on windowed mode.

Specified how to access log files.

Removed the "optional" from screenshots and logs, as a lot of reports end up needing them for more information.

Removed spaces before colons.